+++ src/main/java/com/x0x0b/piyo_todo_api/domain/AttachmentBase.java
@@ -1,7 +1,7 @@
 package com.x0x0b.piyo_todo_api.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.x0x0b.piyo_todo_api.enums.AttachmentType;
+import com.x0x0b.piyo_todo_api.domain.enums.AttachmentType;
 import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.Setter;
