+++ src/main/java/com/x0x0b/piyo_todo_api/domain/AttachmentS3.java
@@ -1,6 +1,6 @@
 package com.x0x0b.piyo_todo_api.domain;
 
-import com.x0x0b.piyo_todo_api.enums.AttachmentType;
+import com.x0x0b.piyo_todo_api.domain.enums.AttachmentType;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
